chore: move syskit/shell_interface into syskit/interface/commands
To prepare for new parts for the interface.
