+++ .rubocop_todo.yml
@@ -99,7 +99,6 @@ Layout/LineLength:
     - 'lib/syskit/scripts/ide.rb'
     - 'lib/syskit/scripts/instanciate.rb'
     - 'lib/syskit/scripts/instanciate_gui.rb'
-    - 'lib/syskit/shell_interface.rb'
     - 'lib/syskit/task_configuration_manager.rb'
     - 'lib/syskit/test/base.rb'
     - 'lib/syskit/test/profile_test.rb'
@@ -172,7 +171,6 @@ Layout/LineLength:
     - 'test/test_properties.rb'
     - 'test/test_property.rb'
     - 'test/test_remote_state_getter.rb'
-    - 'test/test_shell_interface.rb'
     - 'test/test_task_configuration_manager.rb'
     - 'test/test_task_context.rb'
 
@@ -212,7 +210,6 @@ Lint/AmbiguousRegexpLiteral:
     - 'test/test_component.rb'
     - 'test/test_data_service.rb'
     - 'test/test_remote_state_getter.rb'
-    - 'test/test_shell_interface.rb'
 
 # Offense count: 114
 # Configuration parameters: AllowSafeAssignment.
@@ -329,7 +326,6 @@ Lint/UnusedBlockArgument:
     - 'lib/syskit/roby_app/plugin.rb'
     - 'lib/syskit/ros/node.rb'
     - 'lib/syskit/runtime/connection_management.rb'
-    - 'lib/syskit/shell_interface.rb'
     - 'test/coordination/test_data_monitoring_table.rb'
     - 'test/coordination/test_models_task_extension.rb'
     - 'test/models/test_specialization_manager.rb'
@@ -495,7 +491,6 @@ Metrics/AbcSize:
     - 'lib/syskit/runtime/update_task_states.rb'
     - 'lib/syskit/scripts/common.rb'
     - 'lib/syskit/scripts/doc.rb'
-    - 'lib/syskit/shell_interface.rb'
     - 'lib/syskit/task_context.rb'
     - 'lib/syskit/test/network_manipulation.rb'
     - 'lib/syskit/test/profile_assertions.rb'
@@ -738,7 +733,6 @@ Naming/ClassAndModuleCamelCase:
 Naming/MemoizedInstanceVariableName:
   Exclude:
     - 'lib/syskit/roby_app/rest_api.rb'
-    - 'lib/syskit/shell_interface.rb'
 
 # Offense count: 21
 # Configuration parameters: EnforcedStyle, IgnoredPatterns.
@@ -903,7 +897,6 @@ Style/Documentation:
     - 'lib/syskit/scripts/common.rb'
     - 'lib/syskit/scripts/connection_log.rb'
     - 'lib/syskit/scripts/doc.rb'
-    - 'lib/syskit/shell_interface.rb'
     - 'lib/syskit/test.rb'
     - 'lib/syskit/test/base.rb'
     - 'lib/syskit/test/component_test.rb'
@@ -1039,7 +1032,6 @@ Style/HashSyntax:
     - 'lib/syskit/scripts/autodetect_interfaces.rb'
     - 'lib/syskit/scripts/common.rb'
     - 'lib/syskit/scripts/doc.rb'
-    - 'lib/syskit/shell_interface.rb'
     - 'test/models/test_composition_specialization.rb'
     - 'test/models/test_deployment_group.rb'
     - 'test/models/test_specialization_manager.rb'
@@ -1108,7 +1100,6 @@ Style/IfUnlessModifier:
     - 'lib/syskit/scripts/browse.rb'
     - 'lib/syskit/scripts/common.rb'
     - 'lib/syskit/scripts/instanciate.rb'
-    - 'lib/syskit/shell_interface.rb'
     - 'lib/syskit/task_configuration_manager.rb'
     - 'lib/syskit/test/base.rb'
     - 'lib/syskit/test/flexmock_extension.rb'
@@ -1118,7 +1109,6 @@ Style/IfUnlessModifier:
     - 'test/runtime/test_connection_management.rb'
     - 'test/test_deployment.rb'
     - 'test/test_remote_state_getter.rb'
-    - 'test/test_shell_interface.rb'
     - 'test/test_task_context.rb'
 
 # Offense count: 3
@@ -1142,7 +1132,6 @@ Style/MultilineBlockChain:
     - 'test/network_generation/test_engine.rb'
     - 'test/runtime/test_connection_management.rb'
     - 'test/test_deployment.rb'
-    - 'test/test_shell_interface.rb'
 
 # Offense count: 15
 # Cop supports --auto-correct.
@@ -1170,7 +1159,6 @@ Style/RedundantBegin:
     - 'lib/syskit/port.rb'
     - 'lib/syskit/roby_app/rest_api.rb'
     - 'lib/syskit/runtime/connection_management.rb'
-    - 'lib/syskit/shell_interface.rb'
     - 'test/coordination/test_task_script.rb'
     - 'test/models/test_composition.rb'
     - 'test/test/test_task_context_test.rb'
