+++ lib/syskit/interface.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+require "syskit/interface/commands"
