+++ lib/syskit/roby_app/register_plugin.rb
@@ -2,6 +2,6 @@
 
 require "syskit"
 Roby::Application.register_plugin("syskit", Syskit::RobyApp::Plugin) do
-    require "syskit/shell_interface"
+    require "syskit/interface"
     Syskit::RobyApp::Plugin.enable
 end
